refactor: rewrite type connectors with hooklike pattern
