+++ internal/entitlement/static/connector.go
@@ -29,7 +29,7 @@ func (c *connector) GetValue(entitlement *entitlement.Entitlement, at time.Time)
 	}, nil
 }
 
-func (c *connector) SetDefaultsAndValidate(model *entitlement.CreateEntitlementInputs) error {
+func (c *connector) BeforeCreate(model *entitlement.CreateEntitlementInputs, feature *productcatalog.Feature) error {
 	model.EntitlementType = entitlement.EntitlementTypeStatic
 
 	if model.MeasureUsageFrom != nil ||
@@ -54,7 +54,7 @@ func (c *connector) SetDefaultsAndValidate(model *entitlement.CreateEntitlementI
 	return nil
 }
 
-func (c *connector) ValidateForFeature(entitlement *entitlement.CreateEntitlementInputs, feature productcatalog.Feature) error {
+func (c *connector) AfterCreate(entitlement *entitlement.Entitlement) error {
 	return nil
 }
